Refactor checksum and handshake modules
